+++ dds_cli/utils.py
@@ -15,6 +15,19 @@ console = rich.console.Console()
 stderr_console = rich.console.Console(stderr=True)
 
 
+def parse_project_errors(errors):
+    """Parse all errors related to projects."""
+    msg = ""
+    if errors:
+        for unique_error in set(errors.values()):
+            msg += unique_error
+            affected_projects = [x for x, y in errors.items() if y == unique_error]
+            for proj in affected_projects:
+                msg += f"\n   - {proj}"
+
+    return msg
+
+
 def multiple_help_text(item):
     """Return help text for option with multiple=True."""
     return f" Use the option multiple times to specify more than one {item} [multiple]"